+++ test/geo/postgis_test.exs
@@ -4,7 +4,7 @@ defmodule Geo.PostGIS.Test do
   setup do
     opts = [hostname: "localhost", 
     username: "postgres", database: "geo_postgrex_test",
-    extensions: [{Geo.PostGIS, library: Geo}]]
+    extensions: [{Geo.PostGIS.Extension, library: Geo}]]
 
     {:ok, pid} = Postgrex.Connection.start_link(opts)
     {:ok, _result} = Postgrex.Connection.query(pid, "DROP TABLE IF EXISTS point_test, linestring_test, polygon_test, multipoint_test, multilinestring_test, multipolygon_test, geometrycollection_test", [])